Improve the selection metaheurisitc + more tests
This new version of the selection metaheuristic slightly improves the previous
one (Commit dfa2c44). In the tests, the standard deviation of the fitness
values reduces slightly, and the average of the fitness values slightly
increases. Not much gain, but the search is a bit more ‘stable’.

How?

The improvement comes from using the new any_two_greater_than predicate rather
than any_greater_than, both of which control whether the search will use the
‘select_above_avg’ (if the predicate is true) or ‘select_better_than_worst’
(when it is false, falling back to normal CES) selection strategies.

Why?

The difference is that if we use any_two_greater_than we will create a mating
pool that at least will contain two different individuals rather than a single
individual in the case of any_greater_than. Having a mating pool with a single
individual can only lead to an identical offspring and this might lead to a
more premature convergence. If we have a mating pool with at least two, then it
is more likely that the new individuals will be different from the parents,
thus reducing the likelyhood of premature convergence.